requirements_old.txt
cloudpickle==1.2.2
imageio==2.5.0
matplotlib==3.1.1
numpy
Pillow==6.2.0
scikit-image==0.15.0
scipy
pytorch==1.0.0
tensorboardX==1.8
torchvision==0.2.2
graphviz==2.40.1
opencv-python
transformers== 4.40.0
tensorflow==2.16.1
torch==1.11.0
torchvision==0.12.0

torch==1.11.0+cu113
torchvision==0.12.0+cu113